Moved last managers to corresponding package
