+++ operator/src/main/java/org/bf2/operator/managers/SecuritySecretManager.java
@@ -1,4 +1,4 @@
-package org.bf2.operator.secrets;
+package org.bf2.operator.managers;
 
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecretBuilder;
@@ -6,7 +6,6 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.strimzi.api.kafka.model.KafkaResources;
 import org.bf2.common.OperandUtils;
-import org.bf2.operator.managers.InformerManager;
 import org.bf2.operator.resources.v1alpha1.ManagedKafka;
 
 import javax.enterprise.context.ApplicationScoped;
